+++ src/components/Pages/LandingPage/LandingPage.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 .wrapper {
   background: $light;
